# parEvalMap