--- a/core/shared/src/main/scala/fs2/Stream.scala
+++ b/core/shared/src/main/scala/fs2/Stream.scala
@@ -2042,13 +2042,13 @@ final class Stream[+F[_], +O] private[fs2] (private[fs2] val underlying: Pull[F,
       maxConcurrent: Int
   )(f: O => F2[O2])(implicit F: Concurrent[F2]): Stream[F2, O2] = {
 
-    def init(ch: Channel[F2, F2[Outcome[F2, Throwable, O2]]], release: F2[Unit]) =
-      Deferred[F2, Outcome[F2, Throwable, O2]].flatTap { value =>
+    def init(ch: Channel[F2, F2[Either[Throwable, O2]]], release: F2[Unit]) =
+      Deferred[F2, Either[Throwable, O2]].flatTap { value =>
         ch.send(release *> value.get)
       }
 
-    def send(v: Deferred[F2, Outcome[F2, Throwable, O2]]) =
-      (el: Outcome[F2, Throwable, O2]) => v.complete(el).void
+    def send(v: Deferred[F2, Either[Throwable, O2]]) =
+      (el: Either[Throwable, O2]) => v.complete(el).void
 
     parEvalMapAction(maxConcurrent, f)((ch, release) => init(ch, release).map(send))
   }
@@ -2070,8 +2070,8 @@ final class Stream[+F[_], +O] private[fs2] (private[fs2] val underlying: Pull[F,
 
     val init = ().pure[F2]
 
-    def send(ch: Channel[F2, F2[Outcome[F2, Throwable, O2]]], release: F2[Unit]) =
-      (el: Outcome[F2, Throwable, O2]) => release <* ch.send(el.pure[F2])
+    def send(ch: Channel[F2, F2[Either[Throwable, O2]]], release: F2[Unit]) =
+      (el: Either[Throwable, O2]) => release <* ch.send(el.pure[F2])
 
     parEvalMapAction(maxConcurrent, f)((ch, release) => init.as(send(ch, release)))
   }
@@ -2081,9 +2081,9 @@ final class Stream[+F[_], +O] private[fs2] (private[fs2] val underlying: Pull[F,
       f: O => F2[O2]
   )(
       initFork: (
-          Channel[F2, F2[Outcome[F2, Throwable, O2]]],
+          Channel[F2, F2[Either[Throwable, O2]]],
           F2[Unit]
-      ) => F2[Outcome[F2, Throwable, O2] => F2[Unit]]
+      ) => F2[Either[Throwable, O2] => F2[Unit]]
   )(implicit F: Concurrent[F2]): Stream[F2, O2] =
     if (maxConcurrent == 1) evalMap(f)
     else {
@@ -2094,7 +2094,7 @@ final class Stream[+F[_], +O] private[fs2] (private[fs2] val underlying: Pull[F,
       val action =
         (
           Semaphore[F2](concurrency.toLong),
-          Channel.bounded[F2, F2[Outcome[F2, Throwable, O2]]](concurrency),
+          Channel.bounded[F2, F2[Either[Throwable, O2]]](concurrency),
           Deferred[F2, Unit],
           Deferred[F2, Unit]
         ).mapN { (semaphore, channel, stop, end) =>
@@ -2111,8 +2111,8 @@ final class Stream[+F[_], +O] private[fs2] (private[fs2] val underlying: Pull[F,
                 Deferred[F2, Unit].flatMap { pushed =>
                   val init = initFork(channel, pushed.complete(()).void)
                   poll(init).onCancel(releaseAndCheckCompletion).flatMap { send =>
-                    val action = f(el).guaranteeCase(send) *> pushed.get
-                    F.start(stop.get.race(action).guarantee(releaseAndCheckCompletion))
+                    val action = f(el).attempt.flatMap(send) *> pushed.get
+                    F.start(stop.get.race(action) *> releaseAndCheckCompletion)
                   }
                 }
             }
@@ -2126,7 +2126,7 @@ final class Stream[+F[_], +O] private[fs2] (private[fs2] val underlying: Pull[F,
                   case _                  => stop.complete(()) *> releaseAndCheckCompletion
                 }
 
-          val foreground = channel.stream.evalMap(_.flatMap(_.embed(F.canceled >> F.never)))
+          val foreground = channel.stream.evalMap(_.rethrow)
           foreground.onFinalize(stop.complete(()) *> end.get).concurrently(background)
         }