Resolves issues #2297, #2686, #2195, #2726