@nikiforo 2021-10-10
roughly shows 5x performance boost in the benchmark

benchmark/ jmh:run -i 5 -wi 3 -f1 -t1 fs2.benchmark.ParEvalMapBenchmark

original
<pre>
[info] ParEvalMapBenchmark.evalMap                10     100  thrpt    5  16597,710 ± 5339,488  ops/s
[info] ParEvalMapBenchmark.evalMap                10   10000  thrpt    5    263,320 ±   11,201  ops/s
[info] ParEvalMapBenchmark.evalMap               100     100  thrpt    5  19980,999 ±  458,301  ops/s
[info] ParEvalMapBenchmark.evalMap               100   10000  thrpt    5    319,355 ±   11,039  ops/s
[info] ParEvalMapBenchmark.parEvalMap10           10     100  thrpt    5    127,434 ±   83,615  ops/s
[info] ParEvalMapBenchmark.parEvalMap10           10   10000  thrpt    5      1,618 ±    0,104  ops/s
[info] ParEvalMapBenchmark.parEvalMap10          100     100  thrpt    5    146,220 ±    9,165  ops/s
[info] ParEvalMapBenchmark.parEvalMap10          100   10000  thrpt    5      1,661 ±    0,166  ops/s
</pre>

current
<pre>
[info] Benchmark                         (chunkSize)  (size)   Mode  Cnt      Score       Error  Units
[info] ParEvalMapBenchmark.evalMap                10     100  thrpt    5  18323,931 ±   322,369  ops/s
[info] ParEvalMapBenchmark.evalMap                10   10000  thrpt    5    262,852 ±    65,935  ops/s
[info] ParEvalMapBenchmark.evalMap               100     100  thrpt    5  15407,024 ± 18788,319  ops/s
[info] ParEvalMapBenchmark.evalMap               100   10000  thrpt    5    252,926 ±     5,626  ops/s
[info] ParEvalMapBenchmark.parEvalMap10           10     100  thrpt    5    680,162 ±    35,416  ops/s
[info] ParEvalMapBenchmark.parEvalMap10           10   10000  thrpt    5      8,157 ±     0,429  ops/s
[info] ParEvalMapBenchmark.parEvalMap10          100     100  thrpt    5    717,905 ±    40,958  ops/s
[info] ParEvalMapBenchmark.parEvalMap10          100   10000  thrpt    5      7,857 ±     0,469  ops/s
</pre>