@diesalbla 2021-11-03
So, I have rerun the benchmarks. I used the command
```
benchmark/ jmh:run -i 10 -wi 5 -f1 -t4 -gc true -jvmArgs "-XX:MaxRecursiveInlineLevel=3 -XX:MaxInlineSize=50" fs2.benchmark.ParEvalMapBenchmark
```

Since in this case we are concerned about `parEvalMap`, I saw it better to run the benchmarks with more threads. Also, since the measures by @nikiforo had a bit too much error, I added more measuring and warmup iterations. 

#### Current Version (main): 

```
Benchmark                                  (chunkSize)  (size)   Mode  Cnt      Score      Error  Units
evalMap                         10     100  thrpt   10  47488.576 ± 3183.148  ops/s
evalMap                         10   10000  thrpt   10    801.994 ±   80.927  ops/s
evalMap                        100     100  thrpt   10  50132.632 ± 7662.933  ops/s
evalMap                        100   10000  thrpt   10    882.255 ±   44.519  ops/s
parEvalMapUnordered10           10     100  thrpt   10   1595.386 ±  172.605  ops/s
parEvalMapUnordered10           10   10000  thrpt   10     18.865 ±    1.011  ops/s
parEvalMapUnordered10          100     100  thrpt   10   1798.733 ±   65.134  ops/s
parEvalMapUnordered10          100   10000  thrpt   10     19.772 ±    0.609  ops/s
```

#### New version (branch)

```
Benchmark                                  (chunkSize)  (size)   Mode  Cnt      Score      Error  Units
evalMap                         10     100  thrpt   10  52744.008 ± 4044.386  ops/s
evalMap                         10   10000  thrpt   10    844.634 ±   37.255  ops/s
evalMap                        100     100  thrpt   10  55705.254 ± 1655.890  ops/s
evalMap                        100   10000  thrpt   10    998.729 ±   36.727  ops/s

parEvalMapUnordered10           10     100  thrpt   10   1435.734 ±  209.149  ops/s
parEvalMapUnordered10           10   10000  thrpt   10     16.003 ±    2.574  ops/s
parEvalMapUnordered10          100     100  thrpt   10   1303.686 ±   56.432  ops/s
parEvalMapUnordered10          100   10000  thrpt   10     12.682 ±    2.126  ops/s
```

These results seem to indicate that the new implementation is faster with the `evalMap` function, but slower with the `parEvalMapUnordered10`.