@nikiforo 2021-11-03
Great, many thanks!
However, I don’t see the results of the optimized function in your benchmark output.
Benchmark function `evalMap` was added as a reference point. In other words, it can be used to adjust different benchmark runs by `evalMap` results. The PR targets `parEvalMap`, which performance is measured by `parEvalMap10` benchmark function.
I’ve also rerun the benchmark that compares new `parEvalMap` with the `main` implementation using your command.
```
benchmark/ jmh:run -i 10 -wi 5 -f1 -t4 -gc true -jvmArgs "-XX:MaxRecursiveInlineLevel=3 -XX:MaxInlineSize=50" fs2.benchmark.ParEvalMapBenchmark
```

Current Version (main e7bcbb14f6dfa436e7c4b704746ab358ee68bd63):
```
[info] ParEvalMapBenchmark.evalMap                         10     100  thrpt   10  44379,438 ± 4398,278  ops/s
[info] ParEvalMapBenchmark.evalMap                         10   10000  thrpt   10    742,395 ±   27,320  ops/s
[info] ParEvalMapBenchmark.evalMap                        100     100  thrpt   10  50553,217 ± 3633,609  ops/s
[info] ParEvalMapBenchmark.evalMap                        100   10000  thrpt   10    962,177 ±   26,027  ops/s
[info] ParEvalMapBenchmark.parEvalMap10                    10     100  thrpt   10    212,248 ±   12,463  ops/s
[info] ParEvalMapBenchmark.parEvalMap10                    10   10000  thrpt   10      2,127 ±    0,137  ops/s
[info] ParEvalMapBenchmark.parEvalMap10                   100     100  thrpt   10    219,562 ±    2,204  ops/s
[info] ParEvalMapBenchmark.parEvalMap10                   100   10000  thrpt   10      2,203 ±    0,043  ops/s
[info] ParEvalMapBenchmark.parEvalMapUnordered10           10     100  thrpt   10   1781,774 ±   22,137  ops/s
[info] ParEvalMapBenchmark.parEvalMapUnordered10           10   10000  thrpt   10     17,999 ±    1,234  ops/s
[info] ParEvalMapBenchmark.parEvalMapUnordered10          100     100  thrpt   10   1794,607 ±   16,252  ops/s
[info] ParEvalMapBenchmark.parEvalMapUnordered10          100   10000  thrpt   10     19,297 ±    0,754  ops/s
```

New version(branch):
```
[info] ParEvalMapBenchmark.evalMap                         10     100  thrpt   10  47426,136 ± 2354,241  ops/s
[info] ParEvalMapBenchmark.evalMap                         10   10000  thrpt   10    735,417 ±   28,730  ops/s
[info] ParEvalMapBenchmark.evalMap                        100     100  thrpt   10  49624,932 ± 2298,322  ops/s
[info] ParEvalMapBenchmark.evalMap                        100   10000  thrpt   10    939,003 ±   52,114  ops/s
[info] ParEvalMapBenchmark.parEvalMap10                    10     100  thrpt   10   1532,008 ±  102,198  ops/s
[info] ParEvalMapBenchmark.parEvalMap10                    10   10000  thrpt   10     17,050 ±    0,490  ops/s
[info] ParEvalMapBenchmark.parEvalMap10                   100     100  thrpt   10   1649,628 ±   53,896  ops/s
[info] ParEvalMapBenchmark.parEvalMap10                   100   10000  thrpt   10     16,495 ±    3,106  ops/s
[info] ParEvalMapBenchmark.parEvalMapUnordered10           10     100  thrpt   10   1353,258 ±   78,360  ops/s
[info] ParEvalMapBenchmark.parEvalMapUnordered10           10   10000  thrpt   10     16,804 ±    1,217  ops/s
[info] ParEvalMapBenchmark.parEvalMapUnordered10          100     100  thrpt   10   1673,306 ±   18,525  ops/s
[info] ParEvalMapBenchmark.parEvalMapUnordered10          100   10000  thrpt   10     15,995 ±    0,490  ops/s
```

The performance improvement of `parEvalMap` using your command is even more pronounced(7x). However, as you've already mentioned we loose some performance in parEvalMapUnordered.