@nikiforo 2021-11-11
Is blocked by https://github.com/typelevel/fs2/issues/2195